+++ src/components/vectris/view/gamecanvas.rs
@@ -6,8 +6,9 @@ pub fn GameCanvas() -> impl IntoView {
   let state = expect_context::<RwSignal<GameState>>();
   let grid = create_read_slice(state, |state| state.grid);
   let Grid {cells_per_row, cells_per_col} = grid.get();
+  
   view! {
-    <section id="game-canvas">
+    <section id="game-canvas" class="">
       {cells_per_row} - {cells_per_col}
     </section>
   }
